requirements.txt
Django==4.2.5
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.9.0
django-import-export==4.4.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
gunicorn==23.0.0
whitenoise==6.11.0
redis==5.0.1
celery==5.3.5
django-allauth==0.56.1
dj-rest-auth==5.0.1
drf-yasg==1.21.7
twilio==9.10.0
razorpay==2.0.0
python-dotenv==1.0.1
openpyxl==3.1.5
tablib==3.9.0
